Rename symlink.js to symlink-vendor-directory.js
