+++ postinstall.js
@@ -1,3 +1,3 @@
 /* eslint-disable flowtype/require-valid-file-annotation */
 
-require('./symlink.js');
+require('./symlink-vendor-directory.js');
