+++ preinstall.js
@@ -4,4 +4,4 @@ if (process.env.npm_execpath.indexOf('npm') === -1) {
     throw new Error('\x1b[31mYou must use "npm install", yarn is not supported\x1b[0m');
 }
 
-require('./symlink.js');
+require('./symlink-vendor-directory.js');
