+++ symlink-vendor-directory.js
